backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
httpx>=0.24.0
pillow>=10.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
asyncpg>=0.28.0
redis>=5.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0

# Payment SDKs
alipay-sdk-python>=3.6.0
wechatpayv3>=1.2.0

# S3 Storage
boto3>=1.28.0
greenlet